Check for posts with the same name posted on the same day.

+++ lib/tasks/post.rake
@@ -5,21 +5,25 @@ namespace :post do
   task :create do
     include Conversable;
 
-    POSTS_PATH     = 'src/_posts'
-    TEMPLATES_PATH = 'lib/templates'
+    posts_path = 'src/_posts'
+    templates_path = 'lib/templates'
 
     pubdate = Time.now
 
     title = ask?('What would you like to title your post?')
 
     filename = pubdate.strftime('%F') + '-' + title.downcase.gsub('’', '').gsub(/[^0-9A-Za-z]/i, '-').gsub(/-+/, '-').chomp('-')
-    filepath = "#{POSTS_PATH}/#{filename}.md"
+    filepath = "#{posts_path}/#{filename}.md"
 
-    File.open(filepath, 'w') do |f|
-      f.write ERB.new(File.read("#{TEMPLATES_PATH}/post.md.erb")).result(binding)
+    if File.exists?(filepath)
+      alert '! You already created a post today with that title. Try to be more unique.'
+    else
+      File.open(filepath, 'w') do |f|
+        f.write ERB.new(File.read("#{templates_path}/post.md.erb")).result(binding)
+      end
+
+      `open #{filepath}`
     end
-
-    `open #{filepath}`
   end
 end
